pyproject.toml
[project]
name = "sigma-17-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Taha Abbas",email = "taha.abbas@ideofuzion.com"}
]
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-environ (>=0.11.2)",
    "djangorestframework-simplejwt (>=5.3.0)",
    "django-cors-headers (>=4.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
